Refactored src files to remove circular dependencies
