+++ src/auth.js
@@ -108,3 +108,25 @@ export const possiblyAddScopeDirective = ({
   }
   return undefined;
 };
+
+export const addDirectiveDeclarations = (typeMap, config) => {
+  // overwrites any provided directive declarations for system directive names
+  typeMap['cypher'] = parse(
+    `directive @cypher(statement: String) on FIELD_DEFINITION`
+  ).definitions[0];
+  typeMap['relation'] = parse(
+    `directive @relation(name: String, direction: _RelationDirections, from: String, to: String) on FIELD_DEFINITION | OBJECT`
+  ).definitions[0];
+  // TODO should we change these system directives to having a '_Neo4j' prefix
+  typeMap['MutationMeta'] = parse(
+    `directive @MutationMeta(relationship: String, from: String, to: String) on FIELD_DEFINITION`
+  ).definitions[0];
+  typeMap['neo4j_ignore'] = parse(
+    `directive @neo4j_ignore on FIELD_DEFINITION`
+  ).definitions[0];
+  typeMap['_RelationDirections'] = parse(
+    `enum _RelationDirections { IN OUT }`
+  ).definitions[0];
+  typeMap = possiblyAddDirectiveDeclarations(typeMap, config);
+  return typeMap;
+};
