+++ src/utils.js
@@ -1,5 +1,4 @@
 import { print, parse } from 'graphql';
-import { possiblyAddDirectiveDeclarations } from './auth';
 import { v1 as neo4j } from 'neo4j-driver';
 import _ from 'lodash';
 import filter from 'lodash/filter';
@@ -592,28 +591,6 @@ export const getRelationName = relationDirective => {
     // FIXME: should we ignore this error to define default behavior?
     throw new Error('No name argument specified on @relation directive');
   }
-};
-
-export const addDirectiveDeclarations = (typeMap, config) => {
-  // overwrites any provided directive declarations for system directive names
-  typeMap['cypher'] = parse(
-    `directive @cypher(statement: String) on FIELD_DEFINITION`
-  ).definitions[0];
-  typeMap['relation'] = parse(
-    `directive @relation(name: String, direction: _RelationDirections, from: String, to: String) on FIELD_DEFINITION | OBJECT`
-  ).definitions[0];
-  // TODO should we change these system directives to having a '_Neo4j' prefix
-  typeMap['MutationMeta'] = parse(
-    `directive @MutationMeta(relationship: String, from: String, to: String) on FIELD_DEFINITION`
-  ).definitions[0];
-  typeMap['neo4j_ignore'] = parse(
-    `directive @neo4j_ignore on FIELD_DEFINITION`
-  ).definitions[0];
-  typeMap['_RelationDirections'] = parse(
-    `enum _RelationDirections { IN OUT }`
-  ).definitions[0];
-  typeMap = possiblyAddDirectiveDeclarations(typeMap, config);
-  return typeMap;
 };
 
 export const getQueryCypherDirective = resolveInfo => {
